chore: Update lunch break duration default value to 30 minutes
This commit updates the default value for the lunch break duration in the user model to 30 minutes. The previous default value was 45 minutes, but it has been changed to align with the desired behavior. The minimum value for the lunch break duration remains at 0 minutes, and the maximum value is set to 90 minutes.