feat: Add form to request a break
This commit adds a form to the Schedule page that allows users to request a break. Users can select the type of break, specify the time and date, and provide a reason for the break. When the form is submitted, a break request is created and sent to the server. This feature improves the user experience by providing a convenient way for users to request breaks.